+++ components/atoms/icons/ClassIcon.vue
@@ -1,15 +1,6 @@
-<template>
-  <nuxt-img
-    :src="`/images/class-icon-${classSlug}.png`"
-    :alt="altText"
-    height="24"
-    width="24"
-  />
-</template>
-
 <script>
-import PropTypes from '@znck/prop-types'
 import { defineComponent } from '@nuxtjs/composition-api'
+import PropTypes from '@znck/prop-types'
 
 export default defineComponent({
   props: {
@@ -18,3 +9,12 @@ export default defineComponent({
   }
 })
 </script>
+
+<template>
+  <nuxt-img
+    :src="`/images/class-icon-${classSlug}.png`"
+    :alt="altText"
+    height="24"
+    width="24"
+  />
+</template>
