+++ stories/Ui/ClassIcon.stories.js
@@ -1,4 +1,4 @@
-import ClassIcon from '~/components/Ui/ClassIcon.vue'
+import ClassIcon from '~/components/atoms/icons/ClassIcon.vue'
 
 export default {
   component: ClassIcon,
@@ -6,7 +6,7 @@ export default {
     classSlug: 'rogue',
     altText: 'Voleur'
   },
-  title: 'UI/ClassIcon'
+  title: 'atoms/icons/ClassIcon'
 }
 
 const Template = (args, { argTypes }) => ({
